chore: Add code folding regions to app.js for better organization
I thought it would be nice for app.js to roleplay as an organized code file. So I made a regex to prepend all existing manually commented code areas such as "App: FriendLog" with #region so that VSCode can fold them.

There were a couple uncommented regions that are now commented with summaries I can only describe as 'I tried'